pyproject.toml
[project]
name = "strands-repo"
version = "0.1.0"
description = "Strands native example: PromptHub, Evaluations, and OTEL Tracing with LangSmith"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents[openai]",
    "strands-agents-tools",
    "boto3",
    "langsmith",
    "python-dotenv",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
    "requests",
    "langchain-core",
]

[project.optional-dependencies]
notebook = ["jupyter", "ipykernel"]

[tool.uv]
dev-dependencies = ["jupyter", "ipykernel"]
